dotenvx-practice/pyproject.toml
[project]
name = "dotenvx-practice"
version = "0.1.0"
description = ""
authors = [
    {name = "Kin Yuunan",email = "yuunan.kin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
